Redirect from warc.webgraph.io/specs to corresponding IIPC spec
As demonstrated in the example below, the canonical URL for the WARC
file format specification is rather long and therefore does not lend
itself to being cleanly referred to in Git commit messages, Javadoc and
other contexts where line length is constrained. This is especially
true when fragments are involved.

Because this project will implement the WARC file format standard, it is
expected that references to the specification will proliferate through
documentation and commit messages. The pages added in this commit
implement a simple JavaScript-based redirection scheme from a reasonably
short url to the longer canonical url. The goal is to fit these URLs
within the 72-chararacter line length conventional in Git commit
messages, and a 90-character limit in Javadoc.

For example, a request for

    https://warc.webgraph.io/specs/warc-1.0/#file-and-record-model

will be redirected to

    http://iipc.github.io/warc-specifications/specifications/warc-format/warc-1.0/#file-and-record-model

These redirects happen via JavaScript for now in order to keep the
implementation simple and able to be served statically via GitHub Pages.

Note that implementing these redirects via CloudFlare page rules was
considered, but page rules unfortunately do not forward URL fragments,
which is the entire purpose of this mechanism.