Rename Reader/Writer type params from T => V
This reflects that Readers and Writers are, fundamentally, readers and
writers of *values* as opposed to primitives like int, char or arrays
thereof.

Technically, a Reader or Writer can read or write any object, not just
immutable and stateless value objects, but users are strongly encouraged
to favor this style whenever possible.